logic almost done
logic is almost done had ot make a new logic for some decimal related issue and digits with 0 in front of then like e.g. 00100. 00100 was getting counted as 64 instead of 100 because 00100 is 100 in octal and 64 in decimal so instead of 49+00100 = 149 i was getting --113--. but now with new approach its solved